Add Windows Agent crash-loop.ps1 script
Related issue https://github.com/rancher/rancher/issues/10662
In order to support auto upgrade in Windows, add a new ps script for download and update binaries from rancher server.

The crash-loop.ps1 script need to be set to be your failure action of your service.

When we upgrading rancher server, we will need to stop Rancher server for a while.
Then rancher-agent will stop because of it.
And crash-loop.ps1 will be run and keep connectting to Rancher server, download the newest binaries into the host and start the service again.