Don't unset userPrincipalName
When a PRTB/CRTB is created with userPrincipalId set, we translate
that to a userId on the backend. Previously, we were unsetting
userPrincipalId on the backend, but if it is unset, the UI has
now way to display details about the associated user. So, we'll
keep it set so that UI can continue to load the principal.