Fix PSP not working in non-default namespaces/projects
When a new namespace is created and assigned to a project and that
project has a pod security policy, the pod security policy is not binding
correctly.  This is because when the service account handler looks to see
if a binding exists for a given service account in a given namespace, it
only looks for the first policy with a matching parent template
annotation and then exits.  This fixes the issue by making it compare the
name and namespace of the subject in the clusterrole and then creating
the role binding it if does not exist.

Issue:
https://github.com/rancher/rancher/issues/11887