Sync RoleTemplate and ClusterRole
When a RoleTemplate's rules are changed in the mgmt cluster, every
user cluster where that RoleTemplate is referenced needs to have the
corresponding ClusterRole updated.

We had logic for this before but it was not working because the handler
was registered as a "ClusterScopedLifecycle", but RoleTemplate is not
a cluster scoped resource, it is global.

To avoid propogating all RoleTemplates to all clusters, this change also
takes the logic for creating a clusterRole out of the RoleTemplate handler.
The clusterRole will be created on the fly when a PRTB/CRTB uses the RT.