use waiting state instead of ready
make world better by not showing error msg when provisioning. At the
time when creating a cluster, set its initial condition to Waiting and
wait for the first healthcheck check to pass to set it to true so that
once the cluster finish provisioning it will show a transition msg
instead error.