Prevent deletion of PSPTs assigned to a cluster/project
This change prevents the deletion of PSPTs assigend to a cluster or
project.  It accomplishes that in 2 ways: 1) by removing the delete link
on the API and 2) by verifying that no projects or clusters have a
rerefence to a given PSPT before allowing it to be deleted, and throwing
and error if that is not the case.

Issues:
https://github.com/rancher/rancher/issues/12367
https://github.com/rancher/rancher/issues/12368