Promote rules from project roles to cluster
Some resources (right now PVs and storageClasses) are globally scoped
but need to be viewable by project members in ordered to be used
by project members. Therefore, the rules for these resources are
defined in project-scoped roles.

This logic creates clusterRoles and clusterRoleBindings for such
situations so that project members can properly use these resources.