Fixing missing removal links
This change fixes the missing remove links for PSPTs when all PSPTs are
queried at once.  Because the formatter was using the ID from the
apiContext argument instread of the resource argument, the formatter was
incorrectly showing PSPTs as being assigned to clusters and projects
that did not have PSPTs assigned.  This change fixes the formatter to use
the resource.ID instead.

Issue:
https://github.com/rancher/rancher/issues/12926