Fix so project-less namespaces have their PSPT removed
This change makes it so that namespaces have their assigned PSPT removed
after they are moved out of a project.  Before, when a namespace was in a
project that had a PSPT assigned to it, and the namespace was moved to
"None", it would continue to have the PSPT associated with its previous
project.

Issue:
https://github.com/rancher/rancher/issues/12833