Role not found fix
This change corrects an incorrect expression that was supposed to check
if a role was not found but was not, and caused an error to appear in the
logs.

It also adds a new cattle-admin role to the default import template so
clusters without cluster-admin can be imported successfully.

Issue:
https://github.com/rancher/rancher/issues/13196
https://github.com/rancher/rancher/issues/13642