Push workload update only when node changes
We are pushing workload updates on node change so endpoints can be
reconciled. Before this fix, we were doing it on every node update -
which is done every 10 seconds by Kubernetes. With this fix, the update
will be made only when node change would have an impact on the
endpoints.