ignore error on create if secret exists
Problems:
There are intermittent errors coming from provisioning showing secret
already exists. The reason to that is we are using a cache lister to
check if it already exists and if not we will create the secrets.Since
the store call is happening multi-time during provisioning there is a
great chance it will raise conflicts.

Solutions:
Check the error and if it is already created then we don't return the
error. We also want to avoid use api calls as it will cause performance
issue.