Dont immediately set namespace to active
The InitialRolesPopulated condition was  getting set to true
prematurely. This fixes that. Also, change the initial value
of the condition from False to Unknown because False is interpertted
as an error state in rancher.

This was previously fix but reintroduced in 2b6779719115fd4ede7ab4ec19eca19c97e4e4cb
as a result of a rebase conflict.