Clear nodetemplate for user if their ID doesn't match the namespace
Because Watch doesn't use the queryOpts to filter as expected, and because we cheat on nodetemplates by just giving everyone list access, any watcher will receive ALL new nodetemplates, instead of just those for that user. This hack uses the formatter to check if the nodetemplate should be visible, and blanks it out if not.