Store templateContent cache on disk
Refactor catalogs to cache charts to disk, and pull chart data on
demand as needed. Files and questions are now injected at the
formatter to reduce content in CRDs, and the old templateContent
collection is deleted. Upgrade and cache cleanup methods have been
implemented to remove old data and ensure the disk does not fill up.

For rancher/rancher/issues/14322