Update test_pod_security_polices
Problem:
Errors are being thrown in Rancher logs due to incorrect PSPT creation.
PSPTPB are not being verified to exist.

Solution:
Update test to create valid PSPs.
Update test_pod_security_policy_del to test that that PSPTPB's
are being created and assigned to the correct project. This applies
the pspt via an action which will retry if the object is modified.
Improve project_actions to use DeepCopy.
https://github.com/rancher/rancher/issues/18267