Fixing RKE edit issue
This fixes a issue with the RKE schema where it was being overwritten by a bad
set of fields coming from the RKE driver.  This was preventing existing RKE
clusters from being edited on the UI because the config was bad.  The fix
works by correcting the `rancherkubernetesengine` kontainer driver resource
name in the `hasStaticSchema` check in the kontainer driver controller so that
it will not attempt to use the rke_driver.go schema and default to the built
in RKE struct.

Issue:
https://github.com/rancher/rancher/issues/18120