Removing import kontainer driver as it is unecessary
This change removes import driver from the kontainer engine drivers this is
because it is unnecessary (import_driver.go is never called by Rancer) and is
causing difficulties on the UI where it has to be filtered from the list of
drivers.  Also contains cleanup logic for previous versions of Rancher to
remove the existing import driver.

Issue:
https://github.com/rancher/rancher/issues/18273