Set version on upgrade
Version is now being set when in spec and applied spec
when available in the in the cluster data. Additionally,
if the cluster driver is EKS, the config comparison only
considers version; EKS only allows version to be updated.
Prior, comparison was failed and error would be returned
because versions were the same. Now, version are set on
upgrade and updating is only triggered when version is
changed. Refactored exponential backoff and error
checking for schema check.