Revert change to nodeDriver schemas
In our v2.1 and earlier releases, we had a bug that caused machine driver
fields of type int to be presented in the Rancher API as string fields.

We fixed this on master, but later realized that it would be a breaking
API change. We could have dealt with just the breaking API change, but this
also exposed a bug that node driver schemas were not automatically updated
when Rancher was upgraded. The combination of fixing the int-to-string bug
but not fixing the update bug is too confusing. So, we are rolling back the
change to match how it was in v2.1.

We will have to fix these two bugs together at a later date.