Update Norman
Norman now has a recognized option for opting out of pagination.
Prior, passing a limit of -1 would set the pagination limit to
the default, 1000. The backend would serve a paginated response
that the UI automatically aggregates. This led multiple api calls,
each of which taking as long to return as a single non-paginated
result. Now, one non-paginated response will be returned when
limit is -1.