Use system account as creator of gdns provider app
Regular user cannot create gdns provider, because it's now launched
as a catalog app, and the chart for this app needs to list clusters.
So creating a system account with right permissions solves this.
The app needs to be launched only in local cluster. So we're using
the system account for local cluster which gets created during server
start if local cluster is enabled. We also provide cluster-owner role
to this system account after creating it from cluster controller, hence
if used as the gdns provider app's creator, it has all right permissions.