Package catalog into build
This adds a new setting system-catalog which changes the source of the catalogs.
The two options bundled and external with a default of external. If set to
bundled the two rancher catalogs will only use the bundled version of the
catalog from the build and will not attempt to fetch from the internet. If
set to bundled the normal catalog will be disable by default but can be turned
back on. This is to stop install of apps when the image may not be available
to install.