Use ClusterName for KubeConfig User's name
Currently the user name is always the same for every config you generate
(`u-xxxxx`), so the kubeconfigs for 2 clusters cannot be merged together.

Instead, Generate kubeconfigs with the User, Context, and Cluster name all
set to the cluster `name`, so that the configs of any two clusters (with
different names...) can be merged together successfully by kubectl.