Fix pipeline repos timeout
Problem:
Fetching repos results in HTTP client timeout when the number is large. The cause is that paginated responses are held and not proceeded until all of them are available.

Solution:
Proceed paginated response body early.
This pr also update the HTTP client timeout to 30s to have proper tolerance to high latency network.