Remove orphaned system account users on server start
A system user account is created for every cluster and project.
2.2.3 and prior versions did not delete these accounts when their clusters
or projects were deleted. 2.2.4 has the fix to delete these system users on
cluster/project deletion.
But the system users left behind from cluster/project deletion prior to 2.2.4
need to be cleaned up. This only needs to be done once on startup.